+++ src/v/raft/tests/failure_injectable_log.h
@@ -131,9 +131,9 @@ public:
     std::optional<model::offset>
       retention_offset(storage::gc_config) const final;
 
-    uint64_t dirty_segment_bytes() const final;
+    ssize_t dirty_segment_bytes() const final;
 
-    uint64_t closed_segment_bytes() const final;
+    ssize_t closed_segment_bytes() const final;
 
 private:
     ss::shared_ptr<storage::log> _underlying_log;
