+++ src/v/storage/log.h
@@ -222,8 +222,8 @@ public:
      */
     virtual std::optional<model::offset> retention_offset(gc_config) const = 0;
 
-    virtual uint64_t dirty_segment_bytes() const = 0;
-    virtual uint64_t closed_segment_bytes() const = 0;
+    virtual ssize_t dirty_segment_bytes() const = 0;
+    virtual ssize_t closed_segment_bytes() const = 0;
 
 private:
     ntp_config _config;
